Separate out generator logic from the multisrc boilerplate itself
