+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/comicake/ComiCakeGenerator.kt
@@ -1,7 +1,7 @@
 package eu.kanade.tachiyomi.multisrc.comicake
 
-import eu.kanade.tachiyomi.multisrc.ThemeSourceData.SingleLang
-import eu.kanade.tachiyomi.multisrc.ThemeSourceGenerator
+import generator.ThemeSourceData.SingleLang
+import generator.ThemeSourceGenerator
 
 class ComiCakeGenerator : ThemeSourceGenerator {
