+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/eromuse/EroMuseGenerator.kt
@@ -1,7 +1,7 @@
 package eu.kanade.tachiyomi.multisrc.eromuse
 
-import eu.kanade.tachiyomi.multisrc.ThemeSourceData.SingleLang
-import eu.kanade.tachiyomi.multisrc.ThemeSourceGenerator
+import generator.ThemeSourceData.SingleLang
+import generator.ThemeSourceGenerator
 
 class EroMuseGenerator : ThemeSourceGenerator {
 
@@ -12,7 +12,13 @@ class EroMuseGenerator : ThemeSourceGenerator {
     override val baseVersionCode: Int = 1
 
     override val sources = listOf(
-        SingleLang("8Muses", "https://comics.8muses.com", "en", className = "EightMuses", isNsfw = true),
+        SingleLang(
+            "8Muses",
+            "https://comics.8muses.com",
+            "en",
+            className = "EightMuses",
+            isNsfw = true
+        ),
         SingleLang("Erofus", "https://www.erofus.com", "en", isNsfw = true)
     )
