+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/fmreader/FMReaderGenerator.kt
@@ -1,8 +1,8 @@
 package eu.kanade.tachiyomi.multisrc.fmreader
 
-import eu.kanade.tachiyomi.multisrc.ThemeSourceData.MultiLang
-import eu.kanade.tachiyomi.multisrc.ThemeSourceData.SingleLang
-import eu.kanade.tachiyomi.multisrc.ThemeSourceGenerator
+import generator.ThemeSourceData.MultiLang
+import generator.ThemeSourceData.SingleLang
+import generator.ThemeSourceGenerator
 
 class FMReaderGenerator : ThemeSourceGenerator {
 
@@ -16,19 +16,44 @@ class FMReaderGenerator : ThemeSourceGenerator {
      *  most likely the fix is to override popularMangaNextPageSelector()   */
 
     override val sources = listOf(
-            SingleLang("18LHPlus", "https://18lhplus.com", "en", className = "EighteenLHPlus"),
-            SingleLang("Epik Manga", "https://www.epikmanga.com", "tr"),
-            SingleLang("HanaScan (RawQQ)", "https://hanascan.com", "ja", className = "HanaScanRawQQ"),
-            SingleLang("HeroScan", "https://heroscan.com", "en"),
-            SingleLang("KissLove", "https://kissaway.net", "ja"),
-            SingleLang("LHTranslation", "https://lhtranslation.net", "en", overrideVersionCode = 1),
-            SingleLang("Manga-TR", "https://manga-tr.com", "tr", className = "MangaTR"),
-            SingleLang("ManhuaScan", "https://manhuascan.com", "en"),
-            SingleLang("Manhwa18", "https://manhwa18.com", "en"),
-            MultiLang("Manhwa18.net", "https://manhwa18.net", listOf("en", "ko"), className = "Manhwa18NetFactory"),
-            SingleLang("ManhwaSmut", "https://manhwasmut.com", "en", overrideVersionCode = 1),
-            SingleLang("RawLH", "https://lovehug.net", "ja"),
-            SingleLang("Say Truyen", "https://saytruyen.com", "vi"),
+        SingleLang(
+            "18LHPlus",
+            "https://18lhplus.com",
+            "en",
+            className = "EighteenLHPlus"
+        ),
+        SingleLang("Epik Manga", "https://www.epikmanga.com", "tr"),
+        SingleLang(
+            "HanaScan (RawQQ)",
+            "https://hanascan.com",
+            "ja",
+            className = "HanaScanRawQQ"
+        ),
+        SingleLang("HeroScan", "https://heroscan.com", "en"),
+        SingleLang("KissLove", "https://kissaway.net", "ja"),
+        SingleLang(
+            "LHTranslation",
+            "https://lhtranslation.net",
+            "en",
+            overrideVersionCode = 1
+        ),
+        SingleLang("Manga-TR", "https://manga-tr.com", "tr", className = "MangaTR"),
+        SingleLang("ManhuaScan", "https://manhuascan.com", "en"),
+        SingleLang("Manhwa18", "https://manhwa18.com", "en"),
+        MultiLang(
+            "Manhwa18.net",
+            "https://manhwa18.net",
+            listOf("en", "ko"),
+            className = "Manhwa18NetFactory"
+        ),
+        SingleLang(
+            "ManhwaSmut",
+            "https://manhwasmut.com",
+            "en",
+            overrideVersionCode = 1
+        ),
+        SingleLang("RawLH", "https://lovehug.net", "ja"),
+        SingleLang("Say Truyen", "https://saytruyen.com", "vi"),
     )
 
     companion object {
