+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/madara/MadaraGenerator.kt
@@ -1,8 +1,8 @@
 package eu.kanade.tachiyomi.multisrc.madara
 
-import eu.kanade.tachiyomi.multisrc.ThemeSourceData.SingleLang
-import eu.kanade.tachiyomi.multisrc.ThemeSourceData.MultiLang
-import eu.kanade.tachiyomi.multisrc.ThemeSourceGenerator
+import generator.ThemeSourceData.MultiLang
+import generator.ThemeSourceData.SingleLang
+import generator.ThemeSourceGenerator
 
 class MadaraGenerator : ThemeSourceGenerator {
